requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==4.2.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.1
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
kombu==5.3.4
multidict==6.0.4
openai==1.30.1
packaging==24.2
pillow==11.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.5
puremagic==1.28
pydantic==2.7.1
pydantic_core==2.18.2
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.0.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.2
